src/mail_handler/requirements.txt
pandas
psycopg2
openpyxl
python-dotenv
azure-storage-blob